Conflicts:
	TLS-Core/src/main/java/de/rub/nds/tlsattacker/core/config/Config.java
	TLS-Core/src/test/java/de/rub/nds/tlsattacker/core/protocol/parser/SerialisationFullTest.java

+++ TLS-Core/src/test/java/de/rub/nds/tlsattacker/core/protocol/parser/extension/PaddingExtensionParserTest.java
@@ -62,7 +62,7 @@ public class PaddingExtensionParserTest {
         message = parser.parse();
 
         assertArrayEquals(ExtensionType.PADDING.getValue(), message.getExtensionType().getValue());
-        assertEquals(extensionLength, (int) message.getExtensionLength().getValue());
+        assertEquals(extensionLength, (long) message.getExtensionLength().getValue());
         assertArrayEquals(extensionPayload, message.getPaddingBytes().getValue());
     }
